+++ system/src/migrations/updates/_module.mjs
@@ -15,3 +15,4 @@ export {default as Update_230916_1} from "./Update_230916_1.mjs";
 export {default as Update_230920_1} from "./Update_230920_1.mjs";
 export {default as Update_230922_1} from "./Update_230922_1.mjs";
 export {default as Update_230923_1} from "./Update_230923_1.mjs";
+export {default as Update_230923_2} from "./Update_230923_2.mjs";